third_party/blink/web_tests/fast/dom/shadow/custom-pseudo-in-selector-api-expected.txt

Test for crbug.com/273960. Vendor prefixed pseudo elements should not cause DOM exception 12 when using querySelector, querySelectorAll and webkitMatchesSelector.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS sandbox.firstChild.webkitMatchesSelector("input[type='search']::-webkit-search-decoration") is false
PASS sandbox.firstChild.querySelector("input[type='search']::-webkit-search-decoration") is null
PASS sandbox.firstChild.querySelectorAll("input[type='search']::-webkit-search-decoration").length is 0
PASS successfullyParsed is true

TEST COMPLETE
